added simple_nn amp fp translation

+++ lammps_interface/tools.py
@@ -1001,7 +1001,7 @@ def make_params_file(elements, etas, rs_s, g4_eta = 4, cutoff = 6.5,
     for element in elements:
         with open('params_{}'.format(element),'w') as f:
             # G2
-            for species in range(1, len(element) + 1):
+            for species in range(1, len(elements) + 1):
                 for eta, Rs in zip(etas, rs_s):
                     f.write('2 {} 0 {} {} {} 0.0\n'.format(species, cutoff,
                                                            np.round(eta, 6), Rs))
@@ -1051,8 +1051,8 @@ def reorganize_simple_nn_derivative(image, dx_dict):
         zero_check = [a == 0 for a in derivatives]
         if zero_check == [True] * len(derivatives):
             zero_keys.append(key)
-    #for key in zero_keys:
-    #    del d[key]
+    for key in zero_keys:
+        del d[key]
     d = dict(d)
     return d
 
@@ -1115,7 +1115,8 @@ def get_hash(atoms):
     return hash
 
 
-def convert_simple_nn_fps(traj, delete_old=False):
+def convert_simple_nn_fps(traj, delete_old=True):
+    from multiprocessing import Pool
     # make the directories
     if not os.path.isdir('./amp-fingerprints.ampdb'):
         os.mkdir('./amp-fingerprints.ampdb')
@@ -1126,6 +1127,7 @@ def convert_simple_nn_fps(traj, delete_old=False):
     if not os.path.isdir('./amp-fingerprint-primes.ampdb/loose'):
         os.mkdir('amp-fingerprint-primes.ampdb/loose')
     # perform the reorganization
+    """
     for i, image in enumerate(traj):
         pic = pickle.load(open('./data/data{}.pickle'.format(i + 1), 'rb'))
         im_hash = get_hash(image)
@@ -1137,8 +1139,25 @@ def convert_simple_nn_fps(traj, delete_old=False):
         del x_der_dict  # free up memory just in case
         if delete_old:  # in case disk space is an issue
             os.remove('./data/data{}.pickle'.format(i + 1))
+    """
+    with Pool(10) as p:
+        l_trajs = list(enumerate(traj))
+        p.map(reorganize, l_trajs)
     if delete_old:
         os.rmdir('./data')
+
+def reorganize(inp, delete_old=True):
+    i, image = inp
+    pic = pickle.load(open('./data/data{}.pickle'.format(i + 1), 'rb'))
+    im_hash = get_hash(image)
+    x_list = reorganize_simple_nn_fp(image, pic['x'])
+    pickle.dump(x_list, open('./amp-fingerprints.ampdb/loose/' + im_hash, 'wb'))
+    del x_list  # free up memory just in case
+    x_der_dict = reorganize_simple_nn_derivative(image, pic['dx'])
+    pickle.dump(x_der_dict, open('./amp-fingerprint-primes.ampdb/loose/' + im_hash, 'wb'))
+    del x_der_dict  # free up memory just in case
+    if delete_old:  # in case disk space is an issue
+        os.remove('./data/data{}.pickle'.format(i + 1))
 
 
 class DummySimple_nn(object):
@@ -1215,11 +1234,13 @@ def make_simple_nn_fps(traj, descriptors, clean_up_directory=True):
     # generate the descriptors
     descriptor.generate()
     
-    # clean the folder of all the junk
-    files = ['simple_nn_input_traj.traj', 'str_list',
-            'pickle_list', 'simple_nn_log']
-    for file in files:
-        os.remove(file)
+    if clean_up_directory:
+        # clean the folder of all the junk
+        files = ['simple_nn_input_traj.traj', 'str_list',
+                 'pickle_list', 'simple_nn_log']
+        files += list(params.values())
+        for file in files:
+            os.remove(file)
 
 def make_amp_descriptors_simple_nn(traj, g2_etas, g2_rs_s, g4_etas, g4_zetas, g4_gammas, cutoff):
     """
@@ -1234,7 +1255,7 @@ def make_amp_descriptors_simple_nn(traj, g2_etas, g2_rs_s, g4_etas, g4_zetas, g4
                        (g2_etas, g2_rs_s, g4_etas, 
                         cutoff, 
                         g4_zetas, g4_gammas),
-                       clean_up_directory=True)
+                        clean_up_directory=True)
     convert_simple_nn_fps(traj, delete_old=True)
 
 
@@ -1527,6 +1548,35 @@ def toggle_continue_simple_nn(cont, filename='input.yaml'):
         change_yaml_line('  continue: true', filename)
     elif cont == 'no':
         change_yaml_line('  continue: false', filename)
+
+def write_fp_code_input(atoms):
+    os.mkdir('fp_calc')
+    os.chdir('./fp_calc')
+
+    with open('cell_martix', 'w') as f:
+        for line in atoms.get_cell():
+            for entry in line:
+                f.write(str(entry) + ' ')
+            f.write('\n')
+
+    with open('positions', 'w') as f:
+        for atom_pos in atoms.get_scaled_positions():
+            for entry in atom_pos:
+                f.write(str(entry) + ' ')
+            f.write('\n')
+
+    with open('elements', 'w') as f:
+        for element in atoms.get_chemical_symbols():
+            f.write(str(element) + '\n')
+
+    with open('element_alist', 'w') as f:
+        elements = list(set(atoms.get_chemical_symbols()))
+        elements.sort()
+        for element in elements:
+            f.write(element + '\n')
+
+    os.chdir('../')
+
 
 def atomic_parity_plot(traj1, traj2):
     """
